style: Posting 함수 가독성 높이기 위한 함수 처리 - Payload 생성, 필드 유효성 검증, 주소 분리, 일하는 요일 convertDays함수를 /util/postin/ 하위에 함수 파일로 처리 - phoneInput의 체크박스 값이 읽어지지 않는 버그 디버깅
ref : #18
